set 修改

+++ websites/code2/studygolang/src/logic/user_test.go
@@ -6,7 +6,7 @@ import (
 )
 
 func TestFindUserInfos(t *testing.T) {
-	usersMap := DefaultUser.FindUserInfos(nil, []int{1, 2, 3})
+	usersMap := DefaultUser.FindUserInfos(nil, []int64{1, 2, 3})
 	if len(usersMap) == 0 {
 		t.Fatal(usersMap)
 	}
